+++ imports/inputTypes/select-checkbox/select-checkbox.js
@@ -1,11 +1,28 @@
-/*jshint esversion: 6 */
+import { Template } from 'meteor/templating'
+import './select-checkbox.html'
 
-import { Template } from 'meteor/templating';
-import './select-checkbox.html';
-import { dsk } from '../../utilities/dsk';
-import { selectedAttsAdjust } from '../../utilities/selectedAttsAdjust.js';
+Template.afCheckboxGroup_materialize.onCreated(() => {
+  const instance = Template.instance()
+  console.log('select checkbox instance', instance)
+})
 
 Template.afCheckboxGroup_materialize.helpers({
-  dsk:      dsk,
-  itemAtts: selectedAttsAdjust,
-});
+  dsk() {
+    const instance = Template.instance()
+    return {
+      'data-schema-key': instance.data.atts['data-schema-key']
+    }
+  },
+  itemAtts(item) {
+    console.log('item', item)
+    const atts = {
+      type: 'checkbox',
+      value: item.value,
+      checked: item.selected,
+      id: item.atts.id+'_'+item._id,
+      name: item.name
+    }
+    console.log('atts', atts)
+    return atts
+  }
+})