+++ imports/utilities/dsk.js
@@ -1,5 +1,3 @@
-/*jshint esversion: 6 */
-
 export const dsk = function () {
   return {
     'data-schema-key': this.atts['data-schema-key']