+++ imports/utilities/selectedAttsAdjust.js
@@ -1,13 +1,11 @@
-/*jshint esversion: 6 */
-
 export const selectedAttsAdjust = function () {
-  const atts = _.clone(this.atts);
+  const atts = _.clone(this.atts)
 
   if (this.selected) {
-    atts.checked = "";
+    atts.checked = "checked"
   }
 
-  atts.id = atts.id + "_" + this._id;
-  delete atts['data-schema-key'];
-  return atts;
-};
+  atts.id = atts.id + "_" + this._id
+  delete atts['data-schema-key']
+  return atts
+}